moved player, board, and game class to seperate model files
